+++ CHANGELOG.md
@@ -61,6 +61,9 @@
 ### `Bean System` enhancements
 - Show bean-type-specific inline navigable icon [#1045](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1045)
 
+### `ImpEx` enhancements
+- Parse and style field value prefix `model://` [#1077](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1077)
+
 ### Fixes
 - Configure SAP CX JavaDoc for OOTB Modules [#1027](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1027) for Issue [1026](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1026)
 - Removed dependency on Kotlin PSI utils methods [#1053](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1053) for Issue [1052](https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/1052)